+++ CacheContext.test.mjs
@@ -2,7 +2,7 @@
 
 import { strictEqual } from "assert";
 import React from "react";
-import ReactDOMServer from "react-dom/server.js";
+import ReactDOMServer from "react-dom/server";
 
 import Cache from "./Cache.mjs";
 import CacheContext from "./CacheContext.mjs";